% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_layers.R
\name{all.equal.LayersR6}
\alias{all.equal.LayersR6}
\title{Test if two LayersR6 objects are equal}
\usage{
\method{all}{equal.LayersR6}(target, current)
}
\arguments{
\item{target}{\R object.}

\item{current}{other \R object, to be compared with \code{target}.}
}
\description{
Test if two LayersR6 objects are equal
}
