% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_anndata.R, R/class_layers.R,
%   R/class_raw.R
\name{all.equal.AnnDataR6}
\alias{all.equal.AnnDataR6}
\alias{all.equal.LayersR6}
\alias{all.equal.RawR6}
\title{Test if two objects objects are equal}
\usage{
\method{all.equal}{AnnDataR6}(target, current, ...)

\method{all.equal}{LayersR6}(target, current, ...)

\method{all.equal}{RawR6}(target, current, ...)
}
\arguments{
\item{target}{\R object.}

\item{current}{other \R object, to be compared with \code{target}.}

\item{...}{further arguments for different methods, notably the
    following two, for numerical comparison:}
}
\description{
Test if two objects objects are equal
}
