% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnData.R
\name{dimnames.AnnDataR6}
\alias{dimnames.AnnDataR6}
\alias{dim.AnnDataR6}
\alias{as.data.frame.AnnDataR6}
\alias{as.matrix.AnnDataR6}
\alias{r_to_py.AnnDataR6}
\alias{[.AnnDataR6}
\title{AnnData Helpers}
\usage{
\method{dimnames}{AnnDataR6}(x)

\method{dim}{AnnDataR6}(x)

\method{as.data.frame}{AnnDataR6}(x, layer = NULL, ...)

\method{as.matrix}{AnnDataR6}(x, layer = NULL, ...)

\method{r_to_py}{AnnDataR6}(x, convert = FALSE)

\method{[}{AnnDataR6}(x, ..., layer = NULL)
}
\arguments{
\item{x}{An AnnData object}

\item{layer}{An AnnData layer. If \code{NULL}, will use \code{ad$X}, otherwise \code{ad$layers[layer]}.}

\item{convert}{?}
}
\description{
AnnData Helpers
}
