% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_version.R
\name{get_version}
\alias{get_version}
\title{get_version}
\usage{
get_version(
  root = ".",
  version_scheme = "guess-next-dev",
  local_scheme = "node-and-date",
  write_to = NULL,
  write_to_template = NULL,
  relative_to = NULL,
 
    tag_regex = "^(?:[\\\\w-]+-)?(?P<version>[vV]?\\\\d+(?:\\\\.\\\\d+){0,2}[^\\\\+]*)(?:\\\\+.*)?$",
  parentdir_prefix_version = NULL,
  fallback_version = NULL,
  fallback_root = ".",
  parse = NULL,
  git_describe_command = NULL
)
}
\arguments{
\item{root}{root}

\item{version_scheme}{version_scheme}

\item{local_scheme}{local_scheme}

\item{write_to}{write_to}

\item{write_to_template}{write_to_template}

\item{relative_to}{relative_to}

\item{tag_regex}{tag_regex}

\item{parentdir_prefix_version}{parentdir_prefix_version}

\item{fallback_version}{fallback_version}

\item{fallback_root}{fallback_root}

\item{parse}{parse}

\item{git_describe_command}{git_describe_command}
}
\description{
If supplied, relative_to should be a file from which root may
}
\details{
be resolved. Typically called by a script or module that is not
in the root of the repository to direct setuptools_scm to the
root of the repository by supplying ``__file__``.
}
