% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_layers.R
\name{names.LayersR6}
\alias{names.LayersR6}
\alias{length.LayersR6}
\alias{r_to_py.LayersR6}
\alias{py_to_r.anndata._core.aligned_mapping.Layers}
\alias{[.LayersR6}
\alias{[<-.LayersR6}
\alias{[[.LayersR6}
\alias{[[<-.LayersR6}
\title{Layers Helpers}
\usage{
\method{names}{LayersR6}(x)

\method{length}{LayersR6}(x)

\method{r_to_py}{LayersR6}(x, convert = FALSE)

\method{py_to_r}{anndata._core.aligned_mapping.Layers}(x)

\method{[}{LayersR6}(x, name)

\method{[}{LayersR6}(x, name) <- value

\method{[[}{LayersR6}(x, name)

\method{[[}{LayersR6}(x, name) <- value
}
\arguments{
\item{x}{An AnnData object.}

\item{convert}{Not used.}

\item{name}{Name of the layer.}

\item{value}{Replacement value.}
}
\description{
Layers Helpers
}
\examples{
\dontrun{
ad <- AnnData(
  X = matrix(c(0, 1, 2, 3, 4, 5), nrow = 2),
  obs = data.frame(group = c("a", "b"), row.names = c("s1", "s2")),
  var = data.frame(type = c(1L, 2L, 3L), row.names = c("var1", "var2", "var3")),
  layers = list(
    spliced = matrix(c(4, 5, 6, 7, 8, 9), nrow = 2),
    unspliced = matrix(c(8, 9, 10, 11, 12, 13), nrow = 2)
  )
)

ad$layers["spliced"]
ad$layers["test"] <- matrix(c(1, 3, 5, 7), nrow = 2)

length(ad$layers)
names(ad$layers)
}
}
