% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read}
\alias{read}
\title{read}
\usage{
read(filename, backed = NULL)
}
\arguments{
\item{filename}{filename}

\item{backed}{backed}
}
\description{
Read \code{.h5ad}-formatted hdf5 file.
}
\details{
\subsection{Parameters}{

filename File name of data file.
backed If \code{'r'}, load \code{~anndata.AnnData} in \code{backed} mode instead of fully loading it into memory (\code{memory} mode). If you want to modify backed attributes of the AnnData object, you need to choose \code{'r+'}.
as_sparse If an array was saved as dense, passing its name here will read it as a sparse_matrix, by chunk of size \code{chunk_size}.
as_sparse_fmt Sparse format class to read elements from \code{as_sparse} in as.
chunk_size Used only when loading sparse dataset that is stored as dense. Loading iterates through chunks of the dataset of this row size until it reads the whole dataset. Higher size means higher memory consumption and higher (to a point) loading speed.
}
}
