% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_csv.R
\name{read_csv}
\alias{read_csv}
\title{read_csv}
\usage{
read_csv(
  filename,
  delimiter = ",",
  first_column_names = NULL,
  dtype = "float32"
)
}
\arguments{
\item{filename}{Data file.}

\item{delimiter}{Delimiter that separates data within text file. If \code{NULL}, will split at arbitrary number of white spaces, which is different from enforcing splitting at single white space \code{' '}.}

\item{first_column_names}{Assume the first column stores row names.}

\item{dtype}{Numpy data type.}
}
\description{
Read \code{.csv} file.
}
\details{
Same as \code{\link[=read_text]{read_text()}} but with default delimiter \code{','}.
}
\examples{
\dontrun{
ad <- read_csv("matrix.csv")
}
}
