% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_excel.R
\name{read_excel}
\alias{read_excel}
\title{read_excel}
\usage{
read_excel(filename, sheet, dtype = "float32")
}
\arguments{
\item{filename}{File name to read from.}

\item{sheet}{Name of sheet in Excel file.}

\item{dtype}{Numpy data type.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}
Read \code{.xlsx} (Excel) file.
}
\details{
Assumes that the first columns stores the row names and the first row the
column names.
}
\section{Superseded}{

This function is superseded. Please use \href{https://anndataR.scverse.org}{anndataR}
for reading and working with \code{AnnData} objects in R.
See \code{vignette("migration_to_anndataR", package = "anndata")} for migration guidance.
}

\examples{
\dontrun{
ad <- read_excel("spreadsheet.xls")
}
}
