% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mtx.R
\name{read_mtx}
\alias{read_mtx}
\title{read_mtx}
\usage{
read_mtx(filename, dtype = "float32")
}
\arguments{
\item{filename}{The filename.}

\item{dtype}{Numpy data type.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}
Read \code{.mtx} file.
}
\section{Superseded}{

This function is superseded. Please use \href{https://anndataR.scverse.org}{anndataR}
for reading and working with \code{AnnData} objects in R.
See \code{vignette("migration_to_anndataR", package = "anndata")} for migration guidance.
}

\examples{
\dontrun{
ad <- read_mtx("matrix.mtx")
}
}
