% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_text.R
\name{read_text}
\alias{read_text}
\title{read_text}
\usage{
read_text(
  filename,
  delimiter = NULL,
  first_column_names = NULL,
  dtype = "float32"
)
}
\arguments{
\item{filename}{Data file, filename or stream.}

\item{delimiter}{Delimiter that separates data within text file.
If \code{NULL}, will split at arbitrary number of white spaces, which is different
from enforcing splitting at single white space \code{' '}.}

\item{first_column_names}{Assume the first column stores row names.}

\item{dtype}{Numpy data type.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}
Read \code{.txt}, \code{.tab}, \code{.data} (text) file.
}
\details{
Same as \code{\link[=read_csv]{read_csv()}} but with default delimiter \code{NULL}.
}
\section{Superseded}{

This function is superseded. Please use \href{https://anndataR.scverse.org}{anndataR}
for reading and working with \code{AnnData} objects in R.
See \code{vignette("migration_to_anndataR", package = "anndata")} for migration guidance.
}

\examples{
\dontrun{
ad <- read_text("matrix.tab")
}
}
