% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_umi_tools.R
\name{read_umi_tools}
\alias{read_umi_tools}
\title{read_umi_tools}
\usage{
read_umi_tools(filename, dtype = "float32")
}
\arguments{
\item{filename}{File name to read from.}

\item{dtype}{Numpy data type.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}
Read a gzipped condensed count matrix from umi_tools.
}
\section{Superseded}{

This function is superseded. Please use \href{https://anndataR.scverse.org}{anndataR}
for reading and working with \code{AnnData} objects in R.
See \code{vignette("migration_to_anndataR", package = "anndata")} for migration guidance.
}

\examples{
\dontrun{
ad <- read_umi_tools("...")
}
}
