% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_zarr.R
\name{read_zarr}
\alias{read_zarr}
\title{Read from a hierarchical Zarr array store.}
\usage{
read_zarr(store)
}
\arguments{
\item{store}{The filename, a MutableMapping, or a Zarr storage class.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}
}
\section{Superseded}{

This function is superseded. Please use \href{https://anndataR.scverse.org}{anndataR}
for reading and working with \code{AnnData} objects in R.
See \code{vignette("migration_to_anndataR", package = "anndata")} for migration guidance.
}

\examples{
\dontrun{
ad <- read_zarr("...")
}
}
