% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_anndata.R
\name{all.equal.AnnDataR6}
\alias{all.equal.AnnDataR6}
\title{Test if two AnnDataR6 objects are equal}
\usage{
\method{all}{equal.AnnDataR6}(target, current)
}
\arguments{
\item{target}{\R object.}

\item{current}{other \R object, to be compared with \code{target}.}
}
\description{
Test if two AnnDataR6 objects are equal
}
