% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnData.R
\name{AnnData}
\alias{AnnData}
\title{AnnData}
\usage{
AnnData(
  X = NULL,
  obs = NULL,
  var = NULL,
  uns = NULL,
  obsm = NULL,
  varm = NULL,
  layers = NULL,
  raw = NULL,
  dtype = "float32",
  shape = NULL,
  filename = NULL,
  filemode = NULL,
  asview = FALSE
)
}
\arguments{
\item{X}{X}

\item{obs}{obs}

\item{var}{var}

\item{uns}{uns}

\item{obsm}{obsm}

\item{varm}{varm}

\item{layers}{layers}

\item{raw}{raw}

\item{dtype}{dtype}

\item{shape}{shape}

\item{filename}{filename}

\item{filemode}{filemode}

\item{asview}{asview}
}
\description{
An annotated data matrix.
}
\details{
`~anndata.AnnData` stores a data matrix `X` together with annotations
of observations `obs` (`obsm`, `obsp`),
variables `var` (`varm`, `varp`),
and unstructured annotations `uns`. .. figure:: https://falexwolf.de/img/scanpy/anndata.svg :width: 350px An `~anndata.AnnData` object `adata` can be sliced like a
`~pandas.DataFrame`,
for instance `adata_subset = adata[:, list_of_variable_names]`.
`~anndata.AnnData`’s basic structure is similar to R’s ExpressionSet
[Huber15]_. If setting an `.h5ad`-formatted HDF5 backing file `.filename`,
data remains on the disk but is automatically loaded into memory if needed.
See this `blog post`_ for more details. .. _blog post: http://falexwolf.de/blog/171223_AnnData_indexing_views_HDF5-backing/ Parameters
----------
X A #observations × #variables data matrix. A view of the data is used if the data type matches, otherwise, a copy is made.
obs Key-indexed one-dimensional observations annotation of length #observations.
var Key-indexed one-dimensional variables annotation of length #variables.
uns Key-indexed unstructured annotation.
obsm Key-indexed multi-dimensional observations annotation of length #observations. If passing a `~numpy.ndarray`, it needs to have a structured datatype.
varm Key-indexed multi-dimensional variables annotation of length #variables. If passing a `~numpy.ndarray`, it needs to have a structured datatype.
layers Key-indexed multi-dimensional arrays aligned to dimensions of `X`.
dtype Data type used for storage.
shape Shape list (#observations, #variables). Can only be provided if `X` is `NULL`.
filename Name of backing file. See `h5py.File`.
filemode Open mode of backing file. See `h5py.File`. See Also
--------
read_h5ad
read_csv
read_excel
read_hdf
read_loom
read_zarr
read_mtx
read_text
read_umi_tools Notes
-----
`~anndata.AnnData` stores observations (samples) of variables/features
in the rows of a matrix.
This is the convention of the modern classics of statistics [Hastie09]_
and machine learning [Murphy12]_,
the convention of dataframes both in R and Python and the established statistics
and machine learning packages in Python (statsmodels_, scikit-learn_). Single dimensional annotations of the observation and variables are stored
in the `obs` and `var` attributes as `~pandas.DataFrame`\ s.
This is intended for metrics calculated over their axes.
Multi-dimensional annotations are stored in `obsm` and `varm`,
which are aligned to the objects observation and variable dimensions respectively.
Square matrices representing graphs are stored in `obsp` and `varp`,
with both of their own dimensions aligned to their associated axis.
Additional measurements across both observations and variables are stored in
`layers`. Indexing into an AnnData object can be performed by relative position
with numeric indices (like pandas’ :meth:`~pandas.DataFrame.iloc`),
or by labels (like :meth:`~pandas.DataFrame.loc`).
To avoid ambiguity with numeric indexing into observations or variables,
indexes of the AnnData object are converted to strings by the constructor. Subsetting an AnnData object by indexing into it will also subset its elements
according to the dimensions they were aligned to.
This means an operation like `adata[list_of_obs, :]` will also subset `obs`,
`obsm`, and `layers`. Subsetting an AnnData object returns a view into the original object,
meaning very little additional memory is used upon subsetting.
This is achieved lazily, meaning that the constituent arrays are subset on access.
Copying a view causes an equivalent “real” AnnData object to be generated.
Attempting to modify a view (at any attribute except X) is handled
in a copy-on-modify manner, meaning the object is initialized in place.
Here’s an example:: batch1 = adata[adata.obs["batch"] == "batch1", :] batch1.obs["value"] = 0 # This makes batch1 a “real” AnnData object At the end of this snippet: `adata` was not modified,
and `batch1` is its own AnnData object with its own data. Similar to Bioconductor’s `ExpressionSet` and :mod:`scipy.sparse` matrices,
subsetting an AnnData object retains the dimensionality of its constituent arrays.
Therefore, unlike with the classes exposed by :mod:`pandas`, :mod:`numpy`,
and `xarray`, there is no concept of a one dimensional AnnData object.
AnnDatas always have two inherent dimensions, `obs` and `var`.
Additionally, maintaining the dimensionality of the AnnData object allows for
consistent handling of :mod:`scipy.sparse` matrices and :mod:`numpy` arrays. .. _statsmodels: http://www.statsmodels.org/stable/index.html
.. _scikit-learn: http://scikit-learn.org/
}
\section{of observations `obs` (`obsm`, `obsp`),}{

variables `var` (`varm`, `varp`), and unstructured annotations `uns`.
}

\section{variables `var` (`varm`, `varp`),}{

and unstructured annotations `uns`.
}

\section{An `~anndata.AnnData` object `adata` can be sliced like a}{

`~pandas.DataFrame`, for instance `adata_subset = adata[:, list_of_variable_names]`. `~anndata.AnnData`’s basic structure is similar to R’s ExpressionSet [Huber15]_. If setting an `.h5ad`-formatted HDF5 backing file `.filename`, data remains on the disk but is automatically loaded into memory if needed. See this `blog post`_ for more details.
}

\section{in the `obs` and `var` attributes as `~pandas.DataFrame`\ s.}{

This is intended for metrics calculated over their axes. Multi-dimensional annotations are stored in `obsm` and `varm`, which are aligned to the objects observation and variable dimensions respectively. Square matrices representing graphs are stored in `obsp` and `varp`, with both of their own dimensions aligned to their associated axis. Additional measurements across both observations and variables are stored in `layers`.
}

\section{Square matrices representing graphs are stored in `obsp` and `varp`,}{

with both of their own dimensions aligned to their associated axis. Additional measurements across both observations and variables are stored in `layers`.
}

\section{At the end of this snippet}{
 `adata` was not modified,:
and `batch1` is its own AnnData object with its own data.
}

