% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{anndata}
\alias{anndata}
\alias{anndata-package}
\title{anndata - Annotated Data}
\description{
An R wrapper for the Python package
\href{https://github.com/theislab/anndata/tree/58886f09b2e387c6389a2de20ed0bc7d20d1b843}{\verb{theislab/anndata@58886f}}.
Provides a scalable way of keeping track of data
and learned annotations.
Used to read from and write to the h5ad file format.
}
\examples{
\dontrun{
ad <- AnnData(
  X = matrix(c(0, 1, 2, 3), nrow = 2, byrow = TRUE),
  obs = data.frame(group = c("a", "b"), row.names = c("s1", "s2")),
  var = data.frame(type = c(1L, 2L), row.names = c("var1", "var2")),
  layers = list(
    spliced = matrix(c(4, 5, 6, 7), nrow = 2, byrow = TRUE),
    unspliced = matrix(c(8, 9, 10, 11), nrow = 2, byrow = TRUE)
  ),
  varm = list(
    ones = matrix(rep(1L, 10), nrow = 2),
    rand = matrix(rnorm(6), nrow = 2),
    zeros = matrix(rep(0L, 10), nrow = 2)
  ),
  uns = list(a = 1, b = 2, c = list(c.a = 3, c.b = 4))
)

value <- matrix(c(1,2,3,4), nrow = 2)
ad$X
ad$X <- value
ad$X

ad$layers
ad$layers["spliced"]
ad$layers["test"] <- value
ad$layers

ad$to_df()
ad$uns
}
}
