% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_excel.R
\name{read_excel}
\alias{read_excel}
\title{read_excel}
\usage{
read_excel(filename, sheet, dtype = "float32")
}
\arguments{
\item{filename}{filename}

\item{sheet}{sheet}

\item{dtype}{dtype}
}
\description{
Read `.xlsx` (Excel) file.
}
\details{
Assumes that the first columns stores the row names and the first row the
column names. Parameters
----------
filename File name to read from.
sheet Name of sheet in Excel file.
}
