% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_excel.R
\name{read_excel}
\alias{read_excel}
\title{read_excel}
\usage{
read_excel(filename, sheet, dtype = "float32")
}
\arguments{
\item{filename}{File name to read from.}

\item{sheet}{Name of sheet in Excel file.}

\item{dtype}{Numpy data type.}
}
\description{
Read \code{.xlsx} (Excel) file.
}
\details{
Assumes that the first columns stores the row names and the first row the
column names.
}
\examples{
\dontrun{
ad <- read_excel("spreadsheet.xls")
}
}
