% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hdf.R
\name{read_hdf}
\alias{read_hdf}
\title{read_hdf}
\usage{
read_hdf(filename, key)
}
\arguments{
\item{filename}{filename}

\item{key}{key}
}
\description{
Read `.h5` (hdf5) file.
}
\details{
Note: Also looks for fields `row_names` and `col_names`. Parameters
----------
filename Filename of data file.
key Name of dataset in the file.
}
