% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hdf.R
\name{read_hdf}
\alias{read_hdf}
\title{read_hdf}
\usage{
read_hdf(filename, key)
}
\arguments{
\item{filename}{Filename of data file.}

\item{key}{Name of dataset in the file.}
}
\description{
Read \code{.h5} (hdf5) file.
}
\details{
Note: Also looks for fields \code{row_names} and \code{col_names}.
}
\examples{
\dontrun{
ad <- read_hdf("file.h5")
}
}
