% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_text.R
\name{read_text}
\alias{read_text}
\title{read_text}
\usage{
read_text(
  filename,
  delimiter = NULL,
  first_column_names = NULL,
  dtype = "float32"
)
}
\arguments{
\item{filename}{filename}

\item{delimiter}{delimiter}

\item{first_column_names}{first_column_names}

\item{dtype}{dtype}
}
\description{
Read `.txt`, `.tab`, `.data` (text) file.
}
\details{
Same as :func:`~anndata.read_csv` but with default delimiter `NULL`. Parameters
----------
filename Data file, filename or stream.
delimiter Delimiter that separates data within text file. If `NULL`, will split at arbitrary number of white spaces, which is different from enforcing splitting at single white space `' '`.
first_column_names Assume the first column stores row names.
dtype Numpy data type.
}
