% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_umi_tools.R
\name{read_umi_tools}
\alias{read_umi_tools}
\title{read_umi_tools}
\usage{
read_umi_tools(filename, dtype = "float32")
}
\arguments{
\item{filename}{filename}

\item{dtype}{dtype}
}
\description{
Read a gzipped condensed count matrix from umi_tools.
}
\details{
Parameters
----------
filename File name to read from.
}
