% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_zarr.R
\name{read_zarr}
\alias{read_zarr}
\title{Read from a hierarchical Zarr array store.}
\usage{
read_zarr(store)
}
\arguments{
\item{store}{The filename, a MutableMapping, or a Zarr storage class.}
}
\description{
Read from a hierarchical Zarr array store.
}
\examples{
\dontrun{
ad <- read_zarr("...")
}
}
